Rewrite safe_stack
